Refactor how the utils are used
Partly to fit with the previous commit about stubbing modules but also
to give more consistency to how modules are named and used.
