+++ src/js/utils/display-name.js
@@ -1,3 +1,15 @@
+import Overlay from 'o-overlay';
+
+const form = `<form id="o-comments-displayname-form" class="o-forms o-forms">
+		<label for="o-comments-displayname-input" class="o-forms__label">Display name</label>
+		<div class="o-comments--displayname-container">
+			<input type="text" class="o-forms__text o-comments-displayname-input" id="o-comments-displayname-input">
+			<button type="submit" class="o-comments-o-buttons--primary">Save</button>
+		</div>
+		<div id="o-comments-displayname-error" class="o-forms__errortext o-comments-displayname-error"></div>
+	</form>
+</form>`;
+
 const isUnique = (displayName) => {
 	const url = `https://comments-api.ft.com/user/displayname/${displayName}`;
 
@@ -16,7 +28,20 @@ const containsInvalidCharacters = (displayName) => {
 		false;
 };
 
-export default (event) => {
+const prompt = () => {
+	const overlay = new Overlay('displayName', {
+		html: form,
+		heading: {
+			title: 'Choose a commenting display name'
+		}
+	});
+
+	overlay.open();
+
+	return overlay;
+};
+
+const validation = (event) => {
 	event.preventDefault();
 
 	return new Promise(resolve => {
@@ -43,4 +68,9 @@ export default (event) => {
 				});
 		}
 	});
+};
+
+export {
+	prompt,
+	validation
 };
