--- src/js/utils/index.js
@@ -1,11 +1,0 @@
-import fetchJsonWebToken from './fetch-json-web-token';
-import * as events from './events';
-import displayNameValidation from './display-name-validation';
-import displayNamePrompt from './display-name-prompt';
-
-export {
-	fetchJsonWebToken,
-	events,
-	displayNamePrompt,
-	displayNameValidation
-};
