+++ test/utils/auth.test.js
@@ -1,7 +1,7 @@
 import proclaim from 'proclaim';
 import sinon from 'sinon/pkg/sinon';
 import fetchMock from 'fetch-mock';
-import * as auth from '../../src/js/utils/fetch-json-web-token';
+import * as auth from '../../src/js/utils/auth';
 
 const sandbox = sinon.createSandbox();
